pyproject.toml
[tool.poetry]
name = "daily-digest"
version = "0.1.0"
description = "RSS feed aggregator for AI-related news and updates"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
license = "MIT"
packages = [
    {include = "src"}
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
feedparser = "^6.0.10"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
python-dateutil = "^2.8.2"
feedgen = "^0.9.0"
pytz = "^2023.3"
httpx = "^0.25.0"
# Using html.parser instead of lxml for Python 3.13 compatibility
html5lib = "^1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 